fastapi-solution/requirements.txt
redis==4.4.2
elasticsearch[async]==7.9.1
fastapi==0.94
orjson==3.8.7
pydantic==1.10.6
python-dotenv==0.21.1
uvicorn==0.22.0
gunicorn==20.1
email-validator==2.0.0.post1
helios-opentelemetry-sdk==1.0.77
uvloop==0.17.0 ; sys_platform != "win32" and implementation_name == "cpython"
backoff==2.2.1
requests==2.31.0
